Added an Abstraction for rippers that only ever download a single file and want to use the bytes progress bar
