+++ src/main/java/com/rarchives/ripme/ripper/rippers/GfycatRipper.java
@@ -9,18 +9,14 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.rarchives.ripme.ripper.AbstractHTMLRipper;
-import com.rarchives.ripme.utils.Utils;
+import com.rarchives.ripme.ripper.AbstractSingleFileRipper;
 import org.jsoup.nodes.Document;
 import org.jsoup.select.Elements;
 
 import com.rarchives.ripme.utils.Http;
 
 
-public class GfycatRipper extends AbstractHTMLRipper {
-
-    private int bytesTotal = 1;
-    private int bytesCompleted = 1;
+public class GfycatRipper extends AbstractSingleFileRipper {
 
     private static final String HOST = "gfycat.com";
 
@@ -109,27 +105,4 @@ public class GfycatRipper extends AbstractHTMLRipper {
         }
         return vidUrl;
     }
-
-    @Override
-    public String getStatusText() {
-        return Utils.getByteStatusText(getCompletionPercentage(), bytesCompleted, bytesTotal);
-    }
-
-    @Override
-    public int getCompletionPercentage() {
-        return (int) (100 * (bytesCompleted / (float) bytesTotal));
-    }
-
-    @Override
-    public void setBytesTotal(int bytes) {
-        this.bytesTotal = bytes;
-    }
-
-    @Override
-    public void setBytesCompleted(int bytes) {
-        this.bytesCompleted = bytes;
-    }
-
-    @Override
-    public boolean useByteProgessBar() {return true;}
 }
